## Switching over to native OpenAPI: - Removed Smithy:   - Removed Smithy models (`model` folder)   - Removed Gradle   - Removed legacy test framework   - Removed build-openapi-specs workflow - Added Editable OpenAPI spec:   - Added multi-file OpenApi spec (`spec` folder)   - Added a tool to merge multi-file spec into single-file spec   - Added build-single-file-specs workflow - Updated Documents ## Adding Request and Response Bodies
The OpenAPI spec has also been updated with the request and response bodies that we extracted from ElasticSearch OpenAPI spec. This is done within the same PR as the native OpenAPI switch because we also used the ElasticSearch OpenAPI spec to map the query/path schemas to the schemas of many components in the bodies.

Signed-off-by: Theo Truong <theotr@amazon.com>